+++ pkg/tcobs/doc.go
@@ -1,0 +1,4 @@
+// Package tcobs is also a helper for testing the target C-code.
+// Each C function gets a Go wrapper which ist tested in appropriate test functions.
+// For some reason inside the tcobs_test.go an 'import "C"' is not possible.
+package tcobs