+++ pkg/tcobs/tcobs.go
@@ -1,6 +1,3 @@
-// Package tcobs is also a helper for testing the target C-code.
-// Each C function gets a Go wrapper which ist tested in appropriate test functions.
-// For some reason inside the tcobs_test.go an 'import "C"' is not possible.
 package tcobs
 
 // #include <stdint.h>
@@ -8,9 +5,7 @@ package tcobs
 // #cgo CFLAGS: -g -Wall -flto
 import "C"
 import (
-	"bytes"
 	"errors"
-	"io"
 	"unsafe"
 )
 
@@ -137,24 +132,6 @@ func Decode(d, in []byte) (n int, e error) {
 	}
 }
 
-// Decoder expects in s a 0-delimited number of TCOBS packages. If a 0-delimiter is found inside s,
-// it tries to decode all bytes until the 0-delimiter as TCOBS package and writes the result in decoded.
-// If s does not contain any 0-delimiter, decoded is nil, after is s and e is nil.
-// If a 0-delimiter was found, inside 'after' all data after the found delimiter.
-// If the TCOBS decoding of the data until the 0-delimiter failed, decoded has len 0, and e is not nil.
-// For details see TCOBSSpecification.md.
-func Decoder(s []byte) (decoded, after []byte, e error) {
-	tcobs, after, found := bytes.Cut(s, []byte{0})
-	if !found {
-		after = s
-		return
-	}
-	decoded = make([]byte, 2*len(tcobs))
-	n, e := Decode(decoded, tcobs)
-	decoded = decoded[len(decoded)-n:]
-	return
-}
-
 // repeatByte returns the value to repeat
 func repeatByte(offset int, in []byte) byte {
 	if offset == 0 { // left byte of Ri is a sigil byte (probably N)
@@ -162,61 +139,4 @@ func repeatByte(offset int, in []byte) byte {
 	} else {
 		return in[len(in)-1]
 	}
-}
-
-// DO NOT USE code below - Work in progress !
-
-//  // DecodeZ decodes a TCOBS frame `in` (until end or 0-delimiter) to `d` and returns as `n` the valid data length from the end in `d`.
-//  // ATTENTION: d is filled from the end! decoded := d[len(d)-n:] is the final result.
-//  // In `m` are returned the processed bytes inside `in`.
-//  // In case of an error n is 0. n can be 0 without having an error.
-//  // For details see TCOBSSpecification.md.
-//  func DecodeZ(d, in []byte) (n, m int, e error) {
-//  	return
-//  }
-
-type decoder struct {
-	r        io.Reader // inner reader
-	rest     []byte    // not processed incomplete buffer
-	restSize int       // valid data inside rest
-	bufSize  int       // internal buffer size
-}
-
-// NewReader creates a decoder instance and returns its address.
-// r will be used as inner reader and size is used as initial size for the inner buffers.
-func NewReader(r io.Reader, size int) *decoder {
-	d := new(decoder)
-	d.r = r
-	d.rest = make([]byte, size)
-	d.bufSize = size
-	return d
-}
-
-func (p *decoder) Read(buffer []byte) (n int, e error) {
-	buf := make([]byte, p.bufSize)     // internal stack allocation
-	scratch := make([]byte, p.bufSize) // internal stack allocation
-	rest := p.rest[:p.restSize]
-	rCnt := copy(buf, rest)
-	nCnt, e := p.r.Read(buf[rCnt:])
-	if e != nil {
-		p.restSize = copy(p.rest, buf[:rCnt+nCnt])
-		return
-	}
-	input := bytes.Split(buf[:rCnt+nCnt], []byte{0})
-
-	for _, x := range input {
-		var cnt int
-		cnt, e = Decode(scratch, x)
-		//  if e != nil {
-		//  	if i == len(input) { // last element is maybe an incomplete package
-		//  	p.rest = x // keep rest for the next Read
-		//  	return
-		//  	}else{ // probably a merge of 3 packets (data disruption)
-		//  		fmt.Printf( "%w, data disruption?\n", e)
-		//  	}
-		//  }
-		decoded := scratch[len(scratch)-cnt:]
-		n += copy(buffer[n:], decoded)
-	}
-	return
 }