+++ pkg/tcobs/tcobs_test.go
@@ -121,26 +121,6 @@ func TestDecode(t *testing.T) {
 	}
 }
 
-// TestDecoder tests the repeated usage of Decoder.
-func TestDecoder(t *testing.T) {
-	i := []byte{0xAA, 0xBB, 0xA2, 0, 4, 0xA1, 0, 6}
-	before, after, err := tcobs.Decoder(i) // panic
-	assert.True(t, err == nil)
-	assert.Equal(t, i[:2], before)
-	assert.Equal(t, i[4:], after)
-
-	before, after, err = tcobs.Decoder(after)
-	assert.True(t, err == nil)
-	assert.Equal(t, i[4:5], before)
-	assert.Equal(t, i[7:], after)
-
-	before, after, err = tcobs.Decoder(after)
-	assert.True(t, err == nil)
-	var exp []uint8
-	assert.Equal(t, exp, before)
-	assert.Equal(t, i[7:], after)
-}
-
 // TestCEncodeDecode12 tests on generated random byte numbers 0xFF, 0x00, 0x01 and 0x02 for random length 0-32767.
 func TestCEncodeDecode12(t *testing.T) {
 	max := 32768