+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # TCOBS Changelog
 
+## v0.2.0
+
+### Added or Changed
+
+* Code in tcobs.go, tcobs.h, tcobs.c is extensively tested (see tcobs_test.go)
+* Code in remove*.* is tested but should not be used.
+* Code in experimental is WIP (work in progress) and NOT stable.
+
 ## v0.1.0
 
 ### Added or Changed